requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.1.0
django-ckeditor-5==0.0.14
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
python-decouple==3.5
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
django-crispy-forms==1.14.0
django-allauth==0.47.0